+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2025-02-06
+
 ### Added
 
 - [connector] The Morio connector is now backend by Vector (instead of Logstash)